refactor: extract custom-field core CSS to styles folder
